requirements.txt
asgiref==3.5.0
cffi==1.15.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
defusedxml==0.7.1
Django==3.2.10
django-allauth==0.43.0
django-betterforms==1.1.4
django-bootstrap==0.2.4
django-bootstrap-form==3.4
django-email-confirmation==0.2
django-extensions==3.1.5
django-filter==21.1
django-material==1.10.0
django-rest-swagger==2.2.0
django-taggit==1.0.0
django-utils-six==2.0
djangorestframework==3.13.1
httplib2==0.20.4
importlib-metadata==4.11.3
itypes==1.2.0
Jinja2==3.0.3
Markdown==3.3.6
MarkupSafe==2.1.1
oauth2==1.9.0.post1
oauthlib==2.0.0
openapi-codec==1.3.2
pkg_resources==0.0.0
psycopg2==2.8.6
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
python-openid==2.2.5
python3-openid==3.2.0
pytz==2016.7
requests==2.11.1
requests-oauthlib==0.7.0
simplejson==3.17.6
six==1.10.0
sqlparse==0.4.2
Unipath==1.1
uritemplate==4.1.1
zipp==3.7.0
